[hotfix] Merge duplicate CommonRpcMessageUtils
